Polygon double & int coordinated
Memory violation when loading. Removing still needs debugging.

+++ src/visualization/vtkWidget.cpp
@@ -581,7 +581,7 @@ void VtkWidget::removeSurfaceNote2DMark(double* point)
 	mCallback2D->removeSurfaceNoteMark(point);
 }
 
-void VtkWidget::removePolygonNote2DMark(std::vector<std::pair<int, int> >* polygon)
+void VtkWidget::removePolygonNote2DMark(std::vector<std::pair<double, double> >* polygon)
 {
 	mCallback2D->removePolygonNoteMark(polygon);
 	//// TO BE TESTED
@@ -597,7 +597,7 @@ void VtkWidget::openSurfaceNote2DMark(double* point)
 	mCallback2D->openSurfaceNoteMark(point);
 }
 
-void VtkWidget::openPolygonNote2DMark(std::vector<std::pair<int, int> >* polygon)
+void VtkWidget::openPolygonNote2DMark(std::vector<std::pair<double, double> >* polygon)
 {
 	mCallback2D->openPolygonNoteMark(polygon);
 	//// TO BE TESTED
@@ -613,7 +613,7 @@ void VtkWidget::loadSurfaceNote2DMark(double* point, const ColorType color, bool
 	mCallback2D->displayLoadSurfaceNote(point, color, isDisplay);
 }
 
-void VtkWidget::loadPolygonNote2DMark(std::vector<std::pair<int, int> >* polygon, const ColorType color, bool isDisplay)
+void VtkWidget::loadPolygonNote2DMark(std::vector<std::pair<double, double> >* polygon, const ColorType color, bool isDisplay)
 {
 	mCallback2D->displayLoadPolygonNote(polygon, color, isDisplay);
 	//// TO BE TESTED
