+++ src/visualization/vtkWidget.h
@@ -302,9 +302,9 @@ public slots:
   void removePointNote2DMark(double* point);
   void openPointNote2DMark(double* point);
   void loadPointNote2DMark(double* point, const ColorType color, bool isDisplay = false);
-  void removePolygonNote2DMark(std::vector<std::pair<int, int> >* polygon);
-  void openPolygonNote2DMark(std::vector<std::pair<int, int> >* polygon);
-  void loadPolygonNote2DMark(std::vector<std::pair<int, int> >* polygon, const ColorType color, bool isDisplay = false);
+  void removePolygonNote2DMark(std::vector<std::pair<double, double> >* polygon);
+  void openPolygonNote2DMark(std::vector<std::pair<double, double> >* polygon);
+  void loadPolygonNote2DMark(std::vector<std::pair<double, double> >* polygon, const ColorType color, bool isDisplay = false);
 
 protected slots:
   void updateIntensityL12(double intensity1, double intensity2);
